Create production-only config file
We need a config file in the repo for sequelize to run migrations on
heroku.  We also want to keep config.json in gitignore, so that people
don't accidentally push their db credentials.

This adds a config.production.json file that will have the required
option for sequelize to run on Heroku - sequelize commands will need to
be run with the --config option and specify the config file location