Move reports/count route up in route list
if reports/:id comes before reports/count, then the app uses reports/:id and parses
the word 'count' as a report id, routing to the show method. By moving the
reports/count route up in the list, it is matched first.

Also added some error logging to the count routes